requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.25.0
snowflake-connector-python>=3.0.0
python-dotenv>=0.19.0 